CLN: refactoring of Plot class to ColorPlot class
