+++ force_wfmanager/ui/__init__.py
@@ -3,7 +3,7 @@ from .contributed_ui.contributed_ui import ContributedUIHandler  # noqa
 from .contributed_ui.i_contributed_ui import IContributedUI  # noqa
 from .contributed_ui.ui_select_modal import UISelectModal  # noqa
 from .contributed_ui.ui_select_modal import UISelectHandler  # noqa
-from .review.plot import BasePlot # noqa
+from .review.color_plot import BasePlot # noqa
 from .review.i_base_plot import IBasePlot # noqa
 from .review.i_data_view import IDataView # noqa
 from .review.data_view import BaseDataView # noqa
