+++ force_wfmanager/ui/review/tests/test_data_view_pane.py
@@ -3,7 +3,7 @@ import unittest
 
 from force_wfmanager.model.analysis_model import AnalysisModel
 from force_wfmanager.ui.review.data_view_pane import DataViewPane
-from force_wfmanager.ui.review.plot import Plot
+from force_wfmanager.ui.review.color_plot import ColorPlot
 
 
 DATA_VIEW_PANE_EDIT_TRAITS_PATH = \
@@ -43,7 +43,7 @@ class TestDataViewPane(unittest.TestCase):
         # This test class doesn't test the application, so there is no
         # plugin discovery. One default plot will be found as it's the only
         # one which is automatically populated.
-        self.assertIn(Plot, self.pane.available_data_views)
+        self.assertIn(ColorPlot, self.pane.available_data_views)
         self.assertEqual(len(self.pane.available_data_views), 1)
 
     def test_data_view_descriptions(self):
@@ -51,6 +51,7 @@ class TestDataViewPane(unittest.TestCase):
         with mock.patch(DATA_VIEW_PANE_EDIT_TRAITS_PATH):
             self.pane.change_view = True
         self.assertIn(
-            "Plot with colormap (force_wfmanager.ui.review.plot.Plot)",
+            "ColorPlot with colormap "
+            "(force_wfmanager.ui.review.color_plot.ColorPlot)",
             self.pane.data_view_descriptions.values()
         )
